FIX : libft final moulinette

+++ ft_memset.c
@@ -6,15 +6,15 @@
 /*   By: nben-yaa <nben-yaa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/22 15:37:41 by supasaitama       #+#    #+#             */
-/*   Updated: 2018/06/25 10:57:43 by nben-yaa         ###   ########.fr       */
+/*   Updated: 2018/06/27 20:06:22 by nben-yaa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stddef.h>
+#include "libft.h"
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*tmp;
 
 	tmp = (unsigned char *)b;
